third_party/blink/web_tests/wpt_internal/bluetooth/service/getCharacteristics/gen-disconnect-called-before-with-uuid.https.html

<!-- Generated by //third_party/blink/web_tests/bluetooth/generate.py -->
<!DOCTYPE html>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/resources/testdriver.js"></script>
<script src="/resources/testdriver-vendor.js"></script>
<script src="/bluetooth/resources/bluetooth-test.js"></script>
<script src="/bluetooth/resources/bluetooth-fake-devices.js"></script>
<script src="/common/gc.js"></script>
<script>
'use strict';
function createDOMException(func) {
  return new DOMException(
      `Failed to execute '${func}' on 'BluetoothRemoteGATTService': ` +
      `GATT Server is disconnected. Cannot retrieve characteristics. ` +
      `(Re)connect first with \`device.gatt.connect\`.`,
      'NetworkError');
}

bluetooth_test(() => {
  return setBluetoothFakeAdapter('HeartRateAdapter')
    .then(() => requestDeviceWithTrustedClick({
      filters: [{services: ['heart_rate']}]}))
    .then(device => device.gatt.connect())
    .then(gatt => {
      return gatt.getPrimaryService('heart_rate')
        .then(heart_rate_service => {
          gatt.disconnect();
          return assert_promise_rejects_with_message(
            heart_rate_service.getCharacteristics('heart_rate_measurement'),
            createDOMException('getCharacteristics'));
        });
    });
}, 'disconnect() called before getCharacteristics. Reject with NetworkError.');

</script>
